v1.0.0

+++ 题库标准算法/1.两数之和/C++/Solution1.0.0.cpp
@@ -3,11 +3,11 @@ public:
     vector<int> twoSum(vector<int>& nums, int target)
     {
         vector<int> result(2);
-        const unsigned short length = nums.size();
+        const unsigned short length = (unsigned short)nums.size();
 
         for(unsigned short i = 0; i < length - 1; i++)
         {
-            for(unsigned short j = i + 1; j < length; j++)
+            for(unsigned short j = (unsigned short)(i + 1); j < length; j++)
             {
                 if(nums[i] + nums[j] == target)
                 {
